requirements.txt
streamlit==1.28.0
requests==2.32.2
email-validator==2.0.0
protobuf>=3.20,<5.0
cloudinary==1.37.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==6.1.1
coverage==7.5.0
validators==0.20.0
supabase==2.0.0 
python-dotenv==1.0.0